cgame: Move aura functions in one file to keep code maintainability, disable unused custom sounds (these ones were printing warning errors, so no BFP model uses these sounds), add 2 forgotten files in cgame.q3asm (bg_weaponconfig and cg_skinconfig) and remove some non-useful code in cgame
