+++ MakefileQVM
@@ -91,7 +91,7 @@ QA_SRC = \
 
 CG_SRC = \
  cg_main $(CGDIR)/cg_syscalls.asm \
- cg_consolecmds cg_draw cg_drawtools cg_effects cg_ents cg_event cg_info \
+ cg_aura cg_consolecmds cg_draw cg_drawtools cg_effects cg_ents cg_event cg_info \
  cg_localents cg_marks cg_particles cg_players cg_playerstate cg_predict cg_scoreboard \
  cg_servercmds cg_skinconfig cg_snapshot cg_trails cg_view cg_weapons \
  bg_weaponconfig bg_slidemove bg_pmove bg_lib bg_misc \
